+++ decidim-dev/lib/decidim/dev/test/rspec_support/component.rb
@@ -103,6 +103,39 @@ module Decidim
                                               .pluck(:decidim_author_id).flatten.compact.uniq
       end
     end
+
+    class CoauthorableDummyResource < ApplicationRecord
+      include ::Decidim::Coauthorable
+      include HasComponent
+    end
+
+    class OfficialAuthorPresenter
+      def name
+        self.class.name
+      end
+
+      def nickname
+        UserBaseEntity.nicknamize(name)
+      end
+
+      def deleted?
+        false
+      end
+
+      def respond_to_missing?
+        true
+      end
+
+      def method_missing(method, *args)
+        if method.to_s.ends_with?("?")
+          false
+        elsif [:avatar_url, :profile_path, :badge, :followers_count].include?(method)
+          ""
+        else
+          super
+        end
+      end
+    end
   end
 end
 
@@ -156,6 +189,14 @@ Decidim.register_component(:dummy) do |component|
     resource.searchable = true
   end
 
+  component.register_resource(:coauthorable_dummy_resource) do |resource|
+    resource.name = :coauthorable_dummy
+    resource.model_class_name = "Decidim::DummyResources::CoauthorableDummyResource"
+    resource.template = "decidim/coauthorabledummy_resource/linked_dummys"
+    resource.actions = %w(foo-coauthorable)
+    resource.searchable = false
+  end
+
   component.register_stat :dummies_count_high, primary: true, priority: Decidim::StatsRegistry::HIGH_PRIORITY do |components, _start_at, _end_at|
     components.count * 10
   end
@@ -191,6 +232,27 @@ RSpec.configure do |config|
           t.integer :decidim_author_id, index: false
           t.string :decidim_author_type, index: false
           t.integer :decidim_user_group_id, index: false
+          t.references :decidim_category, index: false
+          t.references :decidim_scope, index: false
+          t.string :reference
+
+          t.timestamps
+        end
+      end
+
+      unless ActiveRecord::Base.connection.data_source_exists?("decidim_dummy_resources_coauthorable_dummy_resources")
+        ActiveRecord::Migration.create_table :decidim_dummy_resources_coauthorable_dummy_resources do |t|
+          t.jsonb :translatable_text
+          t.string :title
+          t.string :body
+          t.text :address
+          t.float :latitude
+          t.float :longitude
+          t.datetime :published_at
+          t.integer :coauthorships_count, null: false, default: 0
+          t.integer :endorsements_count, null: false, default: 0
+
+          t.references :decidim_component, index: false
           t.references :decidim_category, index: false
           t.references :decidim_scope, index: false
           t.string :reference